commented out email file/will delete it later, put email function into hangout_router so matched will get email
